% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_cov.R
\name{cor_cov}
\alias{cor_cov}
\title{Correlation and Covariance}
\usage{
cor_cov(x, y)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector of same length}
}
\value{
A list with covariance and correlation
}
\description{
Computes covariance and correlation between two numeric vectors.
}
