% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_exp.R
\name{mod_exp}
\alias{mod_exp}
\title{Modular Exponentiation
Computes (base^exp) mod mod efficiently.}
\usage{
mod_exp(base, exp, mod)
}
\arguments{
\item{base}{Integer base}

\item{exp}{Integer exponent}

\item{mod}{Integer modulus}
}
\value{
Integer result of (base^exp) mod
}
\description{
Modular Exponentiation
Computes (base^exp) mod mod efficiently.
}
