% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_utils.R
\name{normal_cdf}
\alias{normal_cdf}
\title{Normal CDF}
\usage{
normal_cdf(x, mean = 0, sd = 1)
}
\arguments{
\item{x}{Numeric value}

\item{mean}{Mean of the distribution (default = 0)}

\item{sd}{Standard deviation (default = 1)}
}
\value{
Probability that a normal random variable <= x
}
\description{
Computes the cumulative distribution function of the normal distribution.
}
