% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_utils.R
\name{normal_pdf}
\alias{normal_pdf}
\title{Normal PDF}
\usage{
normal_pdf(x, mean = 0, sd = 1)
}
\arguments{
\item{x}{Numeric value}

\item{mean}{Mean of the distribution (default = 0)}

\item{sd}{Standard deviation (default = 1)}
}
\value{
Numeric density value
}
\description{
Computes the probability density function of the normal distribution.
}
