% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime_sieve.R
\name{prime_sieve}
\alias{prime_sieve}
\title{Generate prime numbers up to n}
\usage{
prime_sieve(n)
}
\arguments{
\item{n}{Integer upper limit}
}
\value{
Integer vector of primes <= n
}
\description{
Generate prime numbers up to n
}
